NMS-12722: Generate samples in converter
