+++ features/newts-repository-converter/src/test/resources/sample-counter.xml
